backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Vector Database
chromadb==0.4.22

# AI/ML
ollama==0.1.6
langchain==0.1.4
langchain-community==0.0.16

# Speech Processing
faster-whisper==0.10.0
piper-tts==1.2.0
soundfile==0.12.1
numpy==1.26.3

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# LiveKit
livekit==1.0.23
livekit-api==1.1.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
pillow==10.2.0
opencv-python==4.9.0.80

# Caching
redis==5.0.1

# Logging
loguru==0.7.2

# Testing (optional for MVP)
pytest==7.4.4
pytest-asyncio==0.23.3
